issue 2721  restore object on a cloud - big code structure refactor
