+++ storage-lifecycle-service/sls/app/cp_api_interface.py
@@ -12,61 +12,52 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 #
-
-from sls.app.cp_api_interface import CloudPipelineDataSource
-
-
-class MockedNotificationRESTApiCloudPipelineDataSource(CloudPipelineDataSource):
-
-    def __init__(self, cp_source):
-        self.cp_source = cp_source
+class CloudPipelineDataSource:
 
     def load_storage(self, datastorage_id):
-        return self.cp_source.load_storage(datastorage_id)
+        pass
 
     def load_available_storages(self):
-        return self.cp_source.load_available_storages()
+        pass
 
     def load_lifecycle_rules_for_storage(self, datastorage_id):
-        return self.cp_source.load_lifecycle_rules_for_storage(datastorage_id)
+        pass
 
     def load_lifecycle_rule(self, datastorage_id, rule_id):
-        return self.cp_source.load_lifecycle_rule(datastorage_id, rule_id)
+        pass
 
     def create_lifecycle_rule_execution(self, datastorage_id, rule_id, execution):
-        return self.cp_source.create_lifecycle_rule_execution(datastorage_id, rule_id, execution)
+        pass
 
     def load_lifecycle_rule_executions(self, datastorage_id, rule_id, path=None, status=None):
-        return self.cp_source.load_lifecycle_rule_executions(datastorage_id, rule_id, path, status)
+        pass
 
     def update_status_lifecycle_rule_execution(self, datastorage_id, execution_id, status):
-        return self.cp_source.update_status_lifecycle_rule_execution(datastorage_id, execution_id, status)
+        pass
 
     def delete_lifecycle_rule_execution(self, datastorage_id, execution_id):
-        return self.cp_source.delete_lifecycle_rule_execution(datastorage_id, execution_id)
+        pass
 
     def send_notification(self, subject, body, to_user, copy_users, parameters):
-        return {
-            "mocked": "true"
-        }
+        pass
 
     def load_role(self, role_id):
-        return self.cp_source.load_role(role_id)
+        pass
 
     def load_role_by_name(self, role_name):
-        return self.cp_source.load_role_by_name(role_name)
+        pass
 
     def load_user_by_name(self, username):
-        return self.cp_source.load_user_by_name(username)
+        pass
 
     def load_user(self, user_id):
-        return self.cp_source.load_user(user_id)
+        pass
 
     def load_preference(self, preference_name):
-        return self.cp_source.load_preference(preference_name)
+        pass
 
     def load_regions(self):
-        return self.cp_source.load_regions()
+        pass
 
     def _load_default_lifecycle_rule_notification(self):
-        return self.cp_source._load_default_lifecycle_rule_notification()
+        pass
