+++ storage-lifecycle-service/sls/app/synchronizer/restoring_synchronizer_impl.py
@@ -1,4 +1,17 @@
-# Copyright 2022 EPAM Systems, Inc. (https://www.epam.com/)
+#  Copyright 2022 EPAM Systems, Inc. (https://www.epam.com/)
+#  #
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#  #
+#     http://www.apache.org/licenses/LICENSE-2.0
+#  #
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+#
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,8 +26,8 @@
 # limitations under the License.
 from collections import OrderedDict
 
-from sls.app.storage_synchronizer import StorageLifecycleSynchronizer
-from sls.model.restore_model import StorageLifecycleRestoreAction
+from sls.app.synchronizer.storage_synchronizer_interface import StorageLifecycleSynchronizer
+from sls.pipelineapi.model.restore_action_model import StorageLifecycleRestoreAction
 
 
 class StorageLifecycleRestoringSynchronizer(StorageLifecycleSynchronizer):
