+++ storage-lifecycle-service/sls/cloud/model/cloud_object_model.py
