+++ storage-lifecycle-service/sls/cloud/s3_cloud.py
@@ -1,0 +1,380 @@
+#  Copyright 2022 EPAM Systems, Inc. (https://www.epam.com/)
+#  #
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#  #
+#     http://www.apache.org/licenses/LICENSE-2.0
+#  #
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+#
+
+import datetime
+import os
+import re
+import urllib
+from time import sleep
+
+import boto3
+from botocore.exceptions import ClientError
+
+from sls.cloud.cloud import StorageOperations
+from sls.cloud.model.cloud_object_model import CloudObject
+
+DESTINATION_STORAGE_CLASS_TAG = 'DESTINATION_STORAGE_CLASS'
+CP_SLC_RULE_NAME_PREFIX = 'CP Storage Lifecycle Rule:'
+
+
+def construct_s3_slc_rule(storage_class):
+    return {
+        'ID': CP_SLC_RULE_NAME_PREFIX + storage_class,
+        'Filter': {
+            'Tag': {
+                'Key': DESTINATION_STORAGE_CLASS_TAG,
+                'Value': storage_class
+            }
+        },
+        'Status': 'Enabled',
+        'Transitions': [
+            {
+                'Days': 1,
+                'StorageClass': storage_class
+            },
+        ],
+        'NoncurrentVersionTransitions': [
+            {
+                'NoncurrentDays': 0,
+                'StorageClass': storage_class
+            },
+        ]
+    }
+
+
+class S3StorageOperations(StorageOperations):
+
+    STANDARD = "STANDARD"
+    GLACIER = "GLACIER"
+    GLACIER_IR = "GLACIER_IR"
+    DEEP_ARCHIVE = "DEEP_ARCHIVE"
+    DELETION = "DELETION"
+
+    EXPEDITED_RESTORE_MODE = "EXPEDITED"
+    STANDARD_RESTORE_MODE = "STANDARD"
+    BULK_RESTORE_MODE = "BULK"
+
+    S3_STORAGE_CLASS_TO_RULE = {
+        GLACIER: construct_s3_slc_rule(GLACIER),
+        DEEP_ARCHIVE: construct_s3_slc_rule(DEEP_ARCHIVE),
+        GLACIER_IR: construct_s3_slc_rule(GLACIER_IR)
+    }
+
+    DELETION_RULE = {
+        'ID': CP_SLC_RULE_NAME_PREFIX + DELETION,
+        'Filter': {
+            'Tag': {
+                'Key': DESTINATION_STORAGE_CLASS_TAG,
+                'Value': DELETION
+            }
+        },
+        'Status': 'Enabled',
+        'Expiration': {
+            'Days': 1
+        },
+        'NoncurrentVersionExpiration': {
+            'NoncurrentDays': 1
+        }
+    }
+
+    def __init__(self, config, logger):
+        self.logger = logger
+        self.config = self._verify_config(config)
+        self.aws_s3_client = boto3.client("s3")
+
+    def prepare_bucket_if_needed(self, bucket):
+        try:
+            existing_slc = self.aws_s3_client.get_bucket_lifecycle_configuration(Bucket=bucket)
+        except ClientError as e:
+            self.logger.log("Cannot load BucketLifecycleConfiguration, seems it doesn't exist. {}".format(e))
+            existing_slc = {'Rules': []}
+
+        cp_lsc_rules = [rule for rule in existing_slc['Rules'] if rule['ID'].startswith(CP_SLC_RULE_NAME_PREFIX)]
+
+        if not cp_lsc_rules:
+            self.logger.log("There are no S3 Lifecycle rules for storage: {}, will create it.".format(bucket))
+            slc_rules = existing_slc['Rules']
+            for rule in self.S3_STORAGE_CLASS_TO_RULE.values():
+                slc_rules.append(rule)
+            slc_rules.append(self.DELETION_RULE)
+            self.aws_s3_client.put_bucket_lifecycle_configuration(
+                Bucket=bucket,
+                LifecycleConfiguration={"Rules": slc_rules})
+        else:
+            self.logger.log("There are already defined S3 Lifecycle rules for storage: {}.".format(bucket))
+
+    def list_objects_by_prefix(self, bucket, prefix, list_versions=False, convert_paths=True):
+        result = []
+        paginator = self.aws_s3_client.get_paginator('list_objects' if not list_versions else 'list_object_versions')
+        page_iterator = paginator.paginate(Bucket=bucket, Prefix=self._path_to_s3_format(prefix))
+        for page in page_iterator:
+            if not list_versions:
+                if 'Contents' in page:
+                    for obj in page['Contents']:
+                        result.append(self._map_s3_obj_to_cloud_obj(obj, convert_paths))
+            else:
+                if 'Versions' in page:
+                    for version in page['Versions']:
+                        result.append(self._map_s3_obj_to_cloud_obj(version, convert_paths))
+        return result
+
+    def tag_files_to_transit(self, bucket, files, storage_class, region, transit_id):
+        tag_operation = {
+            'S3PutObjectTagging': {
+                'TagSet': [
+                    {
+                        'Key': DESTINATION_STORAGE_CLASS_TAG,
+                        'Value': storage_class
+                    },
+                ]
+            }
+        }
+        job_description = self._run_s3_batch_operation(bucket, files, region, tag_operation, transit_id)
+        return self._is_s3_batch_operation_succeeded(job_description)
+
+    def run_files_restore(self, bucket, files, days, restore_tear, region, restore_operation_id):
+        restore_operation = {
+            'S3InitiateRestoreObject': {
+                'ExpirationInDays': days,
+                'GlacierJobTier': restore_tear
+            }
+        }
+        archived_files_to_restore = [f for f in files if f.storage_class != self.STANDARD]
+        self.logger.log("Filtered '{}' actual archived files to restore.".format(len(files)))
+        job_description = self._run_s3_batch_operation(
+            bucket, archived_files_to_restore, region, restore_operation, restore_operation_id)
+        return {
+            "status": self._is_s3_batch_operation_succeeded(job_description),
+            "reason": job_description,
+            "value": None
+        }
+
+    def check_files_restore(self, bucket, files, updated, restore_mode):
+        is_restore_possibly_ready = self._check_if_restore_could_be_ready(next(iter(files), None), updated, restore_mode)
+        restored_till_value = None
+        if not is_restore_possibly_ready:
+            self.logger.log("Probably files is steel restoring because appropriate period of time is not passed. "
+                            "Will not send head request for each file. Last update time: {}. restore mode: {}.".format(updated, restore_mode))
+            return {
+                "status": False,
+                "value": None,
+                "reason": ""
+            }
+        else:
+            self.logger.log("Files may be ready, so will check each file. "
+                            "Last update time: {}. restore mode: {}.".format(updated, restore_mode))
+            for file in files:
+                if file.version_id:
+                    result = self.aws_s3_client.head_object(Bucket=bucket, Key=file.path, VersionId=file.version_id)
+                else:
+                    result = self.aws_s3_client.head_object(Bucket=bucket, Key=file.path)
+                if not result:
+                    self.logger.log(
+                        "Cannot get head of object: '{}' in bucket: {}"
+                        .format(file.path, bucket))
+                    return {
+                        "status": False,
+                        "value": None,
+                        "reason": "Cannot get head for: '{}' bucket: {}, or 'Restore' property is not set"
+                        .format(file.path, bucket)
+                    }
+                else:
+                    if "Restore" not in result:
+                        continue
+
+                    restoring_status_match = re.match('.*expiry-date=["\'](.*)["\']', result["Restore"])
+                    if not restoring_status_match:
+                        return {
+                            "status": False,
+                            "value": None,
+                            "reason": "Status is {} for file: '{}' bucket: {}"
+                            .format(result["Restore"], file.path, bucket)
+                        }
+                    else:
+                        file_restored_till_value = datetime.datetime.strptime(
+                            restoring_status_match.group(1), "%a, %d %b %Y %H:%M:%S %Z")
+                        if restored_till_value is None or file_restored_till_value < restored_till_value:
+                            restored_till_value = file_restored_till_value
+
+            return {
+                "status": True,
+                "value": restored_till_value.strftime("%Y-%m-%d %H:%M:%S.000"),
+                "reason": "All files are ready!"
+            }
+
+    def _run_s3_batch_operation(self, bucket, files, region, operation_obj, operation_id):
+        aws_s3control_client = boto3.client("s3control", region_name=region)
+        is_files_versioned = True if next(iter(files), None).version_id else False
+        manifest_content = "\n".join(
+            ["{},{}".format(bucket, urllib.parse.quote(self._path_to_s3_format(file.path)))
+             if not is_files_versioned
+             else
+             "{},{},{}".format(bucket, urllib.parse.quote(self._path_to_s3_format(file.path)), file.version_id)
+             for file in files]
+        )
+
+        job_location_prefix = os.path.join(self.config["tagging_job_report_bucket_prefix"], bucket, operation_id)
+
+        manifest_key = os.path.join(
+            job_location_prefix,
+            "_".join([bucket, operation_id,
+                      str(datetime.datetime.utcnow()), ".csv"]).replace(" ", "_").replace(":", "_").replace("/", ".")
+        )
+        manifest_object = self.aws_s3_client.put_object(
+            Body=manifest_content,
+            Bucket=self.config["tagging_job_report_bucket"],
+            Key=manifest_key,
+            ServerSideEncryption='AES256'
+        )
+
+        s3_tagging_job = aws_s3control_client.create_job(
+            AccountId=self.config["tagging_job_aws_account_id"],
+            ConfirmationRequired=False,
+            Operation=operation_obj,
+            Report={
+                'Bucket': "arn:aws:s3:::" + self.config["tagging_job_report_bucket"],
+                'Format': 'Report_CSV_20180820',
+                'Enabled': True,
+                'Prefix': job_location_prefix,
+                'ReportScope': 'FailedTasksOnly'
+            },
+            Manifest={
+                'Spec': {
+                    'Format': 'S3BatchOperations_CSV_20180820',
+                    'Fields': ['Bucket', 'Key', 'VersionId'] if is_files_versioned else ['Bucket', 'Key']
+                },
+                'Location': {
+                    'ObjectArn': "".join(
+                        ["arn:aws:s3:::", self.config["tagging_job_report_bucket"], "/", manifest_key]),
+                    'ETag': manifest_object["ETag"]
+                }
+            },
+            Priority=1,
+            RoleArn=self.config["tagging_job_role_arn"]
+        )
+
+        s3_tagging_job_description = None
+        for try_i in range(self.config["tagging_job_poll_status_retry_count"]):
+            self.logger.log("Get Job {} status with try: {}".format(s3_tagging_job["JobId"], try_i))
+            s3_tagging_job_description = aws_s3control_client.describe_job(
+                AccountId=self.config["tagging_job_aws_account_id"],
+                JobId=s3_tagging_job["JobId"]
+            )
+            if "Job" in s3_tagging_job_description and "Status" in s3_tagging_job_description["Job"]:
+                if s3_tagging_job_description["Job"]["Status"] == "Complete" or s3_tagging_job_description["Job"]["Status"] == "Failed":
+                    self.logger.log("Job status: {}. Proceeding.".format(s3_tagging_job_description["Job"]["Status"]))
+                    break
+                else:
+                    self.logger.log("Job status: {}. Wait.".format(s3_tagging_job_description["Job"]["Status"]))
+            sleep(self.config["tagging_job_poll_status_sleep_sec"])
+
+        if not self._is_s3_batch_operation_succeeded(s3_tagging_job_description):
+            self.logger.log(
+                "Job status is not successful. Will keep manifest and report and skip. Task summary: {}".format(
+                    str(s3_tagging_job_description))
+            )
+        else:
+            self._clean_up_after_job(job_location_prefix, manifest_key, s3_tagging_job["JobId"])
+
+        return s3_tagging_job_description
+
+    def _clean_up_after_job(self, job_location_prefix, manifest_key, job_id):
+        job_report_dir = os.path.join(job_location_prefix, "job-" + job_id)
+        job_related_files = self.list_objects_by_prefix(self.config["tagging_job_report_bucket"], job_report_dir,
+                                                        list_versions=False, convert_paths=False)
+        for file in job_related_files:
+            self.aws_s3_client.delete_object(
+                Bucket=self.config["tagging_job_report_bucket"],
+                Key=file.path
+            )
+        self.aws_s3_client.delete_object(
+            Bucket=self.config["tagging_job_report_bucket"],
+            Key=manifest_key
+        )
+
+    @staticmethod
+    def _is_s3_batch_operation_succeeded(s3_tagging_job_description):
+        return s3_tagging_job_description and "Job" in s3_tagging_job_description and \
+               "Status" in s3_tagging_job_description["Job"] \
+               and s3_tagging_job_description["Job"]["Status"] == "Complete" \
+               and "ProgressSummary" in s3_tagging_job_description["Job"] \
+               and s3_tagging_job_description["Job"]["ProgressSummary"]["NumberOfTasksFailed"] == 0
+
+    @staticmethod
+    def _map_s3_obj_to_cloud_obj(s3_object, convert_path):
+        return CloudObject(
+            S3StorageOperations._path_from_s3_format(s3_object["Key"]) if convert_path else s3_object["Key"],
+            s3_object["LastModified"],
+            s3_object["StorageClass"],
+            s3_object["VersionId"] if 'VersionId' in s3_object else None
+        )
+
+    @staticmethod
+    def _verify_config(config):
+        result = dict(config)
+        if "tagging_job_aws_account_id" not in result:
+            raise RuntimeError("Please provide tagging_job_aws_account_id within S3 storage lifecycle service cloud configuration")
+
+        if "tagging_job_report_bucket" not in result:
+            raise RuntimeError("Please provide tagging_job_report_bucket within S3 storage lifecycle service cloud configuration")
+
+        if "tagging_job_role_arn" not in result:
+            raise RuntimeError("Please provide tagging_job_role_arn within S3 storage lifecycle service cloud configuration")
+
+        if "tagging_job_report_bucket_prefix" not in result:
+            result["tagging_job_report_bucket_prefix"] = "cp_storage_lifecycle_tagging"
+        result["tagging_job_report_bucket_prefix"] = result.get("tagging_job_report_bucket_prefix").strip("/")
+
+        if "tagging_job_poll_status_retry_count" not in result:
+            result["tagging_job_poll_status_retry_count"] = 30
+        if result["tagging_job_poll_status_retry_count"] < 1:
+            raise RuntimeError(
+                "Value tagging_job_poll_status_retry_count within S3 storage lifecycle service cloud configuration should be > 1")
+
+        if "tagging_job_poll_status_sleep_sec" not in result:
+            result["tagging_job_poll_status_sleep_sec"] = 5
+        if result["tagging_job_poll_status_sleep_sec"] < 1:
+            raise RuntimeError(
+                "Value tagging_job_poll_status_sleep_sec within S3 storage lifecycle service cloud configuration should be > 1")
+
+        return result
+
+    @staticmethod
+    def _path_from_s3_format(path):
+        return "/" + path if not path.startswith("/") else path
+
+    @staticmethod
+    def _path_to_s3_format(path):
+        return path.replace("/", "", 1) if path.startswith("/") else path
+
+    # For more information on periods see:
+    # https://docs.aws.amazon.com/AmazonS3/latest/userguide/restoring-objects-retrieval-options.html
+    def _check_if_restore_could_be_ready(self, file, updated, restore_mode):
+        if not file or not updated:
+            return False
+        now = datetime.datetime.now(datetime.timezone.utc)
+        restore_mode = self.STANDARD_RESTORE_MODE if not restore_mode else restore_mode
+        check_shift_period = datetime.timedelta(hours=12)
+        if file.storage_class == self.GLACIER or file.storage_class == self.GLACIER_IR:
+            if restore_mode == self.BULK_RESTORE_MODE:
+                check_shift_period = datetime.timedelta(hours=12)
+            elif restore_mode == self.STANDARD_RESTORE_MODE:
+                check_shift_period = datetime.timedelta(hours=5)
+        if file.storage_class == self.DEEP_ARCHIVE:
+            if restore_mode == self.BULK_RESTORE_MODE:
+                check_shift_period = datetime.timedelta(hours=48)
+            elif restore_mode == self.STANDARD_RESTORE_MODE:
+                check_shift_period = datetime.timedelta(hours=12)
+        return updated + check_shift_period < now
